correct working but confusing use of R1 to reset INT0
Change R1 to R0 in the STR that resets INT0

R0 is set to R0, then R1 is used as source register for an STR. this does work - because R1 is 0 at this time without being explicitly set to 0.

Storing the contents of R0 instead of R1 makes it clear that 0 will be stored.